Create  hello.cpp
C++ programs have a very specific structure in terms of how the code is written. There are some key elements that you use in all your C++ programs.

Here, we have a program called hello.cpp. It is a classic first program! This program writes the phrase “Hello, World!” to your terminal.